+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.5.4
 
 * Fix ``focus`` method on select
+* Table with ``selected`` prop can now be natigated with keyboard
 * #251, #345 Add ``bottom-left`` slot to table
 * #334 Add ``defaultDialogConfirmText`` and ``defaultDialogCancelText`` constructor options
 * #335 Add ``icon`` and ``iconPack`` prop to dialog